Database/Gameobjects : Set Phasemask to uint32

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -2043,7 +2043,7 @@ void ObjectMgr::LoadGameobjects()
         if (data.spawnMask & ~spawnMasks[data.mapid])
             sLog->outError(LOG_FILTER_SQL, "Table `gameobject` has gameobject (GUID: %u Entry: %u) that has wrong spawn mask %u including not supported difficulty modes for map (Id: %u), skip", guid, data.id, data.spawnMask, data.mapid);
 
-        data.phaseMask      = fields[18].GetUInt16();
+        data.phaseMask      = fields[18].GetUInt32();
         int16 gameEvent     = fields[19].GetInt8();
         uint32 PoolId        = fields[20].GetUInt32();
